+++ src/pages/configurePatch/configurePatchCore/ParametersContent.tsx
